+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 Version 0.1
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Initial version of rpm2json utility
+- Searched repo directory and reads information from all RPMs
+- Writes one summary JSON file
+- For each RPM found writes a information JSON file